+++ packages/screenshoter/CHANGELOG.md
@@ -3,8 +3,12 @@
 
 ## Unreleased
 
+
+## 3.3.11 - 2022/3/12
+
 - disable scroll to 0,0 before taking stitched screenshot
 - add better error handling when taking a stitched screenshot
+- updated to @applitools/snippets@2.2.1 (from 2.2.0)
 
 ## 3.3.10 - 2022/2/16